Fix awful read() bug in libeti_worker.cc
I somehow thought that EAGAIN was a negative number and that I could
check the return value of read() for this, but that is not how it works.
Instead what was happening is any packets that are 11 bytes long were
ignored, creating very hard to reproduce bugs in libeti services.